feat: Move and restructure VxPersistenceHandler
